requirements.txt
python-dotenv>=0.19.0
requests>=2.27.1
Pillow>=9.0.0
urllib3>=1.26.8
flask>=2.0.0
werkzeug>=2.0.0 